requirements.txt
tensorflow
tensorflow-hub
tensorflow-datasets
numpy
matplotlib
Pillow
